wpa_supplicant: Add get_capability freq
Add "get_capability freq" command to print a more verbose list of channels and
frequencies in MHz. The tag "NO_IBSS" is added, if IBSS mode is not allowed on
a channel. This allows userspace programs to get the frequencies and
limitations associated with each channel.

This extends the information already provided in "get_capability channel" but a
new interface is chosen because of backwards compatibility considerations.

Signed-hostap: Bruno Randolf <br1@einfach.org>
Change-Id: Ic05a080b3cdea3ab095127d8c2e86eb16408bb4f